Changed the Connection Method
IP address is now provided in the client rather than a command line argument. Regex pattern matching is used to determine if the submitted string is appropriately formatted. Options include: An IP address or "localhost", I have not added functionality to the regex to look for a hostname although that is relatively trivial.